backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
chromadb
pypdf
sentence-transformers
python-dotenv
